+++ test/Spec.hs
@@ -12,7 +12,7 @@ import qualified Canon.CanonicalCommentTest
 import qualified Canon.DecisionsTest
 import qualified Canon.IgnoreTest
 import qualified Canon.ProjectTest
-import qualified Canon.Extract.Antlr4Test
+import qualified Canon.Extract.GrammarTest
 import qualified Canon.Model.CheckTest
 import qualified Canon.Git.ParseTest
 import qualified Canon.Model.YamlTest
@@ -25,7 +25,7 @@ main =
   defaultMain
     ( testGroup
         "canon"
-        [ testGroup "unit" [CanonTest.tests, Canon.Antlr4.VendoredTest.tests, Canon.Extract.Antlr4Test.tests]
+        [ testGroup "unit" [CanonTest.tests, Canon.Antlr4.VendoredTest.tests, Canon.Extract.GrammarTest.tests]
         , testGroup
             "property"
             [ Canon.Antlr4.EscapeTest.tests